requirements.txt
auth==0.5.3
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
dnspython==1.16.0
eventlet==0.30.2
falcon==3.0.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
mongoengine==0.23.0
moviepy==1.0.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
proglog==0.1.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
